Add isActive to ndelius sentence model
