+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/ndelius/Sentence.kt
@@ -5,10 +5,12 @@ import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.Sentence as Integ
 
 data class Sentence(
   val date: String? = null,
+  val active: Boolean? = null,
 ) {
   fun toSentence(): IntegrationApiSentence {
     return IntegrationApiSentence(
       dateOfSentencing = LocalDate.parse(this.date),
+      isActive = this.active,
     )
   }
 }
